+++ Assets/_Scripts/Units/Player/Components/PlayerCollisions.cs
@@ -13,19 +13,10 @@ namespace Metroidvania.Player
         /// <summary>Is the player touching the ground?</summary>
         public bool isGrounded { get; private set; }
 
-        /// <summary>Is the player left hand touching a wall?</summary>
-        public bool isTouchingLeftWall { get; private set; }
-
-        /// <summary>Is the player right hand touching a wall?</summary>
-        public bool isTouchingRightWall { get; private set; }
-
         public bool canStand { get; private set; }
 
-        /// <summary>Is the player touching a ledge?</summary>
-        public bool isTouchingLedge { get; private set; }
-
         /// <summary>Is the player touching a wall?</summary>
-        public bool isTouchingWall => isTouchingLeftWall || isTouchingRightWall;
+        public bool isTouchingWall { get; private set; }
 
         public PlayerCollisions(PlayerController player) : base(player)
         {
@@ -47,12 +38,8 @@ namespace Metroidvania.Player
             var position = (Vector2)transform.localPosition;
 
             isGrounded = OverlapBoxOnGround(collisionDataSource.feetRect);
-            isTouchingLeftWall = OverlapBoxOnGround(collisionDataSource.leftHandRect);
-            isTouchingRightWall = OverlapBoxOnGround(collisionDataSource.rightHandRect);
+            isTouchingWall = OverlapBoxOnGround(collisionDataSource.handRect);
             canStand = !OverlapBoxOnGround(player.data.crouchHeadRect);
-
-            // isTouchingLedge = Physics2D.Raycast(position + collisionDataSource.ledgeCheckOffset,
-            //     new Vector2(collisionDataSource.ledgeCheckLength * localScale.x, 0), 0, player.data.wallLayer);
         }
 
         public void SetCollisionsData(PlayerDataChannel.ColliderData colliderData)
@@ -65,7 +52,9 @@ namespace Metroidvania.Player
 
         public Collider2D OverlapBoxOnGround(Rect bounds)
         {
-            return Physics2D.OverlapBox((Vector2)transform.position + bounds.position * (Vector2)transform.localScale, bounds.size, 0, player.data.groundLayer);
+            var playerPosition = (Vector2)transform.position;
+            var boundsPosition = bounds.position * (Vector2)transform.localScale;
+            return Physics2D.OverlapBox(playerPosition + boundsPosition, bounds.size, 0, player.data.groundLayer);
         }
     }
 }